Move TracingOptions/ITracingOptions into Common, so it can be shared between Management.Tracing and Logging.DynamicSerilog.
