+++ src/Common/src/Common/Properties/AssemblyInfo.cs
@@ -7,3 +7,5 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Steeltoe.Common.Test")]
 [assembly: InternalsVisibleTo("Steeltoe.Common.Hosting")]
 [assembly: InternalsVisibleTo("Steeltoe.Common.Hosting.Test")]
+[assembly: InternalsVisibleTo("Steeltoe.Management.Tracing")]
+[assembly: InternalsVisibleTo("Steeltoe.Management.Tracing.Test")]
