+++ src/Management/src/Tracing/TracingLogProcessor.cs
@@ -4,8 +4,8 @@
 
 using System.Text;
 using OpenTelemetry.Trace;
+using Steeltoe.Common;
 using Steeltoe.Extensions.Logging;
-using Steeltoe.Management.OpenTelemetry.Trace;
 
 namespace Steeltoe.Management.Tracing;
